+++ lib/src/database/message/additional/payload.dart
@@ -1,0 +1,98 @@
+import 'package:bson/bson.dart' show BsonBinary, BsonCString, BsonMap;
+
+abstract class Payload {
+  void packValue(BsonBinary buffer);
+
+  int get byteLength;
+
+  Map<String, Object> get content;
+}
+
+class Payload0 extends Payload {
+  BsonMap document;
+
+  Payload0(Map<String, Object> document) : document = BsonMap(document);
+
+  Payload0.fromBuffer(BsonBinary buffer)
+      : document = BsonMap(<String, Object>{})
+          ..unpackValue(buffer..makeByteList());
+
+  @override
+  packValue(BsonBinary buffer) => document.packValue(buffer);
+
+  @override
+  int get byteLength => document.byteLength();
+
+  @override
+  Map<String, Object> get content => document.data;
+}
+
+class Payload1 extends Payload {
+  int _length;
+  final BsonCString identifier;
+  List<BsonMap> _documents;
+
+  Payload1(String identifier, List<Map<String, Object>> documents)
+      : identifier = BsonCString(identifier),
+        _documents = _createBsonMapList(documents);
+
+  Payload1.fromBuffer(BsonBinary buffer)
+      : _length = (buffer..makeByteList()).readInt32(),
+        identifier = BsonCString(buffer.readCString()) {
+    _documents =
+        _decodeBsonMapList(buffer, _length - 4 - identifier.byteLength());
+  }
+
+  @override
+  int get byteLength => _length ??=
+      4 /* sequence length */ + identifier.byteLength() + documentsByteLength;
+
+  int get documentsByteLength {
+    int len = 0;
+    for (BsonMap doc in _documents) {
+      len += doc.byteLength();
+    }
+    return len;
+  }
+
+  @override
+  packValue(BsonBinary buffer) {
+    buffer.writeInt(byteLength);
+    identifier.packValue(buffer);
+    for (BsonMap data in _documents) {
+      data.packValue(buffer);
+    }
+  }
+
+  @override
+  Map<String, Object> get content =>
+      {identifier.data: _extractBsonMapList(_documents)};
+}
+
+List<BsonMap> _createBsonMapList(List<Map<String, Object>> documents) {
+  List<BsonMap> _documents = <BsonMap>[];
+  for (var document in documents) {
+    _documents.add(BsonMap(document));
+  }
+  return _documents;
+}
+
+List<Map<String, Object>> _extractBsonMapList(List<BsonMap> documents) {
+  List<Map<String, Object>> _documents = <Map<String, Object>>[];
+  for (var document in documents) {
+    _documents.add(document.data);
+  }
+  return _documents;
+}
+
+List<BsonMap> _decodeBsonMapList(BsonBinary buffer, int length) {
+  List<BsonMap> _documents = <BsonMap>[];
+  while (length > 0) {
+    BsonMap map = BsonMap(<String, Object>{});
+    map.unpackValue(buffer);
+    _documents.add(map);
+    length -= map.byteLength();
+  }
+
+  return _documents;
+}