+++ lib/src/network/connection.dart
@@ -8,6 +8,7 @@ class _ServerCapabilities {
   bool listCollections = false;
   bool listIndexes = false;
   int maxNumberOfDocsInBatch = 1000;
+  bool supportsOpMsg = false;
 
   getParamsFromIstMaster(Map<String, dynamic> isMaster) {
     if (isMaster.containsKey('maxWireVersion')) {
@@ -24,6 +25,9 @@ class _ServerCapabilities {
       listCollections = true;
       listIndexes = true;
     }
+    if (maxWireVersion >= 6) {
+      supportsOpMsg = true;
+    }
   }
 }
 
@@ -33,17 +37,21 @@ class _Connection {
   ServerConfig serverConfig;
   Socket socket;
   Set<int> _pendingQueries = Set();
-  Map<int, Completer<MongoReplyMessage>> get _replyCompleters =>
+
+  Map<int, Completer<MongoResponseMessage>> get _replyCompleters =>
       _manager.replyCompleters;
+
   Queue<MongoMessage> get _sendQueue => _manager.sendQueue;
-  StreamSubscription<MongoReplyMessage> _repliesSubscription;
-  StreamSubscription<MongoReplyMessage> get repliesSubscription =>
+  StreamSubscription<MongoResponseMessage> _repliesSubscription;
+
+  StreamSubscription<MongoResponseMessage> get repliesSubscription =>
       _repliesSubscription;
 
   bool connected = false;
   bool _closed = false;
   bool isMaster = false;
   final _ServerCapabilities serverCapabilities = _ServerCapabilities();
+  final ServerStatus serverStatus = ServerStatus();
 
   _Connection(this._manager, [this.serverConfig]) {
     if (serverConfig == null) {
@@ -51,43 +59,52 @@ class _Connection {
     }
   }
 
-  Future<bool> connect() {
-    Completer<bool> completer = Completer();
-    Socket.connect(serverConfig.host, serverConfig.port).then((Socket _socket) {
-      // Socket connected.
-      socket = _socket;
-      _repliesSubscription =
-          MongoMessageHandler().transformer.bind(socket).listen(_receiveReply,
-              onError: (e, st) {
-                _log.severe("Socket error ${e} ${st}");
-                //completer.completeError(e);
-                if (!_closed) {
-                  _onSocketError();
-                }
-              },
-              cancelOnError: true,
-              onDone: () {
-                if (!_closed) {
-                  _onSocketError();
-                }
-              });
-      connected = true;
-      completer.complete(true);
-    }).catchError((err) {
-      completer.completeError(err);
-    });
-    return completer.future;
+  Future<bool> connect() async {
+    Socket _socket;
+    try {
+      _socket = await Socket.connect(serverConfig.host, serverConfig.port);
+    } catch (e, st) {
+      _log.severe("Socket error on connect(): ${e} ${st}");
+      _closed = true;
+      connected = false;
+      var ex = const ConnectionException("Could not connect to the Data Base.");
+      throw ex;
+    }
+
+    // ignore: unawaited_futures
+    _socket.done.catchError((error) => _log.info("Socket error ${error}"));
+    socket = _socket;
+
+    _repliesSubscription = socket
+        .transform<MongoResponseMessage /*MongoReplyMessage*/ >(
+            MongoMessageHandler().transformer)
+        .listen(_receiveReply,
+            onError: (e, st) {
+              _log.severe("Socket error ${e} ${st}");
+              if (!_closed) {
+                _onSocketError();
+              }
+            },
+            cancelOnError: true,
+            onDone: () {
+              if (!_closed) {
+                _onSocketError();
+              }
+            });
+    connected = true;
+    return true;
   }
 
   Future close() {
     _closed = true;
+    connected = false;
     return socket.close();
   }
 
   _sendBuffer() {
-    _log.fine(() => '_sendBuffer ${!_sendQueue.isEmpty}');
+    _log.fine(() => '_sendBuffer ${_sendQueue.isNotEmpty}');
     List<int> message = [];
-    while (!_sendQueue.isEmpty) {
+    while (_sendQueue.isNotEmpty) {
       var mongoMessage = _sendQueue.removeFirst();
       message.addAll(mongoMessage.serialize().byteList);
     }
@@ -126,7 +143,23 @@ class _Connection {
     }
   }
 
-  void _receiveReply(MongoReplyMessage reply) {
+  Future<MongoModernMessage> executeModernMessage(
+      MongoModernMessage modernMessage) {
+    Completer<MongoModernMessage> completer = Completer();
+    if (!_closed) {
+      _replyCompleters[modernMessage.requestId] = completer;
+      _pendingQueries.add(modernMessage.requestId);
+      _log.fine(() => 'Message $modernMessage');
+      _sendQueue.addLast(modernMessage);
+      _sendBuffer();
+    } else {
+      completer.completeError(const ConnectionException(
+          "Invalid state: Connection already closed."));
+    }
+    return completer.future;
+  }
+
+  void _receiveReply(MongoResponseMessage reply) {
     _log.fine(() => reply.toString());
     Completer completer = _replyCompleters.remove(reply.responseTo);
     _pendingQueries.remove(reply.responseTo);
@@ -142,12 +175,16 @@ class _Connection {
 
   void _onSocketError() {
     _closed = true;
+    connected = false;
     var ex = const ConnectionException("connection closed.");
     _pendingQueries.forEach((id) {
       Completer completer = _replyCompleters.remove(id);
       completer.completeError(ex);
     });
     _pendingQueries.clear();
+    if (isMaster) {
+      _manager.close();
+    }
   }
 }